refactor(labs): Move theming to labs
