+++ modules/_labs/core/react/index.ts
@@ -1,6 +1,10 @@
 import type from './lib/type';
 import space from './lib/space';
+import CanvasProvider from './lib/CanvasProvider';
+import {breakpoints, CanvasBreakpoints, BreakpointKey} from './lib/theming/breakpoints';
 
 export default type;
-export {type, space};
+export {breakpoints, type, space, BreakpointKey, CanvasBreakpoints, CanvasProvider};
 export * from './lib/type';
+export * from './lib/theming/types';
+export * from './lib/theming/theme';
